move fst_gwss to parent class
