+++ malariagen_data/af1.py
@@ -10,8 +10,6 @@ except ImportError:
 
 import malariagen_data  # used for .__version__
 
-from .util import CacheMiss
-
 MAJOR_VERSION_INT = 1
 MAJOR_VERSION_GCS_STR = "v1.0"
 CONFIG_PATH = "v1.0-config.json"
@@ -105,6 +103,7 @@ class Af1(AnophelesDataResource):
     _gcs_url = GCS_URL
     _pca_results_cache_name = PCA_RESULTS_CACHE_NAME
     _snp_allele_counts_results_cache_name = SNP_ALLELE_COUNTS_CACHE_NAME
+    _fst_gwss_results_cache_name = FST_GWSS_CACHE_NAME
     _default_site_mask = DEFAULT_SITE_MASK
     _site_annotations_zarr_path = SITE_ANNOTATIONS_ZARR_PATH
     _cohorts_analysis = None
@@ -303,232 +302,3 @@ class Af1(AnophelesDataResource):
         # See also https://github.com/malariagen/malariagen-data-python/issues/306
 
         return super().genome_features(region=region, attributes=attributes)
-
-    # TODO: here plot_haplotype_network
-
-    def fst_gwss(
-        self,
-        contig,
-        window_size,
-        cohort1_query,
-        cohort2_query,
-        sample_sets=None,
-        site_mask="funestus",
-        cohort_size=30,
-        random_seed=42,
-    ):
-        """Run a Fst genome-wide scan to investigate genetic differentiation
-        between two cohorts.
-
-        Parameters
-        ----------
-        contig: str
-            Chromosome arm (e.g., "2RL")
-        window_size : int
-            The size of windows used to calculate h12 over.
-        cohort1_query : str
-            A pandas query string which will be evaluated against the sample
-            metadata e.g., "taxon == 'funestus' and country == 'Ghana'".
-        cohort2_query : str
-            A pandas query string which will be evaluated against the sample
-            metadata e.g., "taxon == 'funestus' and country == 'Ghana'".
-        sample_sets : str or list of str, optional
-            Can be a sample set identifier (e.g., "1229-VO-GH-DADZIE-VMF00095") or a list of
-            sample set identifiers (e.g., ["1240-VO-CD-KOEKEMOER-VMF00099", "1240-VO-MZ-KOEKEMOER-VMF00101"]) or a
-            release identifier (e.g., "1.0") or a list of release identifiers.
-        site_mask : str, optional
-            Site filters mask to apply.
-        cohort_size : int, optional
-            If provided, randomly down-sample to the given cohort size.
-        random_seed : int, optional
-            Random seed used for down-sampling.
-
-        Returns
-        -------
-        x : numpy.ndarray
-            An array containing the window centre point genomic positions.
-        fst : numpy.ndarray
-            An array with Fst statistic values for each window.
-        """
-
-        # change this name if you ever change the behaviour of this function, to
-        # invalidate any previously cached data
-        name = FST_GWSS_CACHE_NAME
-
-        params = dict(
-            contig=contig,
-            window_size=window_size,
-            cohort1_query=cohort1_query,
-            cohort2_query=cohort2_query,
-            sample_sets=self._prep_sample_sets_arg(sample_sets=sample_sets),
-            site_mask=site_mask,
-            cohort_size=cohort_size,
-            random_seed=random_seed,
-        )
-
-        try:
-            results = self.results_cache_get(name=name, params=params)
-
-        except CacheMiss:
-            results = self._fst_gwss(**params)
-            self.results_cache_set(name=name, params=params, results=results)
-
-        x = results["x"]
-        fst = results["fst"]
-
-        return x, fst
-
-    def plot_fst_gwss_track(
-        self,
-        contig,
-        window_size,
-        cohort1_query,
-        cohort2_query,
-        sample_sets=None,
-        site_mask="funestus",
-        cohort_size=30,
-        random_seed=42,
-        title=None,
-        width=DEFAULT_GENOME_PLOT_WIDTH,
-        height=200,
-        show=True,
-        x_range=None,
-    ):
-        """Run and plot a Fst genome-wide scan to investigate genetic
-            differentiation between two cohorts.
-        Parameters
-        ----------
-        contig: str
-            Contig name (e.g., "2RL")
-        window_size : int
-            The size of windows used to calculate h12 over.
-        cohort1_query : str
-            A pandas query string which will be evaluated against the sample
-            metadata e.g., "taxon == 'funestus' and country == 'Ghana'".
-        cohort2_query : str
-            A pandas query string which will be evaluated against the sample
-            metadata e.g., "taxon == 'funestus' and country == 'Ghana'".
-        sample_sets : str or list of str, optional
-            Can be a sample set identifier (e.g., "1229-VO-GH-DADZIE-VMF00095") or a list of
-            sample set identifiers (e.g., ["1240-VO-CD-KOEKEMOER-VMF00099", "1240-VO-MZ-KOEKEMOER-VMF00101"]) or a
-            release identifier (e.g., "1.0") or a list of release identifiers.
-        site_mask : {"funestus"}, optional
-            Site filters mask to apply.
-        cohort_size : int, optional
-            If provided, randomly down-sample to the given cohort size.
-        random_seed : int, optional
-            Random seed used for down-sampling.
-        title : str, optional
-            If provided, title string is used to label plot.
-        width : int, optional
-            Plot width in pixels (px).
-        height : int. optional
-            Plot height in pixels (px).
-        show : bool, optional
-            If True, show the plot.
-        x_range : bokeh.models.Range1d, optional
-            X axis range (for linking to other tracks).
-
-        Returns
-        -------
-        fig : figure
-            A plot showing windowed Fst statistic across chosen contig.
-        """
-
-        # Here we override the superclass implementation in order to provide a
-        # different default value for the `site_mask` parameter.
-        #
-        # Also, we take the opportunity to customise the docstring to use
-        # examples specific to funestus.
-
-        return super().plot_fst_gwss_track(
-            contig=contig,
-            window_size=window_size,
-            cohort1_query=cohort1_query,
-            cohort2_query=cohort2_query,
-            sample_sets=sample_sets,
-            site_mask=site_mask,
-            cohort_size=cohort_size,
-            random_seed=random_seed,
-            title=title,
-            width=width,
-            height=height,
-            show=show,
-            x_range=x_range,
-        )
-
-    def plot_fst_gwss(
-        self,
-        contig,
-        window_size,
-        cohort1_query,
-        cohort2_query,
-        sample_sets=None,
-        site_mask="funestus",
-        cohort_size=30,
-        random_seed=42,
-        title=None,
-        width=DEFAULT_GENOME_PLOT_WIDTH,
-        track_height=190,
-        genes_height=DEFAULT_GENES_TRACK_HEIGHT,
-    ):
-        """Run and plot a Fst genome-wide scan to investigate genetic
-        differentiation between two cohorts.
-
-        Parameters
-        ----------
-        contig: str
-            Contig name (e.g., "2RL")
-        window_size : int
-            The size of windows used to calculate h12 over.
-        cohort1_query : str
-            A pandas query string which will be evaluated against the sample
-            metadata e.g., "taxon == 'funestus' and country == 'Ghana'".
-        cohort2_query : str
-            A pandas query string which will be evaluated against the sample
-            metadata e.g., "taxon == 'funestus' and country == 'Ghana'".
-        sample_sets : str or list of str, optional
-            Can be a sample set identifier (e.g., "1229-VO-GH-DADZIE-VMF00095") or a list of
-            sample set identifiers (e.g., ["1240-VO-CD-KOEKEMOER-VMF00099", "1240-VO-MZ-KOEKEMOER-VMF00101"]) or a
-            release identifier (e.g., "1.0") or a list of release identifiers.
-        site_mask : {"funestus"}, optional
-            Site filters mask to apply.
-        cohort_size : int, optional
-            If provided, randomly down-sample to the given cohort size.
-        random_seed : int, optional
-            Random seed used for down-sampling.
-        title : str, optional
-            If provided, title string is used to label plot.
-        width : int, optional
-            Plot width in pixels (px).
-        track_height : int. optional
-            GWSS track height in pixels (px).
-        genes_height : int. optional
-            Gene track height in pixels (px).
-
-        Returns
-        -------
-        fig : figure
-            A plot showing windowed Fst statistic with gene track on x-axis.
-        """
-
-        # Here we override the superclass implementation in order to provide a
-        # different default value for the `site_mask` parameter.
-        #
-        # Also, we take the opportunity to customise the docstring to use
-        # examples specific to funestus.
-
-        return super().plot_fst_gwss(
-            contig=contig,
-            window_size=window_size,
-            cohort1_query=cohort1_query,
-            cohort2_query=cohort2_query,
-            sample_sets=sample_sets,
-            site_mask=site_mask,
-            cohort_size=cohort_size,
-            random_seed=random_seed,
-            title=title,
-            width=width,
-            track_height=track_height,
-            genes_height=genes_height,
-        )
